Implements my_bipar_checker adds test_graphs/
-> test_graphs/ contains graphs, that may be loaded using LEDAs
test_graph function. They are there to demonstrate that my
implementation of the bipartite checker and LEDAs, both return false if
the graph is not bipartite, but they may return a different odd circle
as a proof.
-> I think this is purely of how I am iterating through the edges, in
order to check if source and target color is the same. Thus I think
there is no point on changing the way I find the odd circle.