Fixes memory leaks originating from DLTree and bugs introduced by circles in graphs.
- Most of the segmentation faults where due to not setting the parent of a Node
to NULL whenever the parent Node was deleted.
- Memory leaks where also introduced because DLTree was not freeing the Nodes
that had created using the new operator, upon desctruction. An alternative
way to do this would be to use shared_ptr for the FORWARD and BACKWARD
matrices, as well as for the children of a Node and weak_ptr for the parent
of a Node and for the _edges list in DLTree.
- Debugging this mess was REALLY painful but I learned ALOT about pointers.
Takeaway messages:
	1. Use BOOST whenever possible. (I am pretty sure I could avoid DLTree that way)
	2. Use shared_ptr, it has an overhead but it will save you a lot of time.
	3. Whenever using new make sure you think about how you are going to delete that allocated memory, before actually creating the memory leaks. Locating those and restructuring your code is a lot harder
TODO: Read about RAII